관광시설 정보 출력 및 필터링 기능 업데이트

+++ app/src/main/java/com/hk/portfolio/today_weather/data/repository/WeatherRepositoryImpl.kt
@@ -25,8 +25,8 @@ class WeatherRepositoryImpl @Inject constructor(
 
     @RequiresApi(Build.VERSION_CODES.O)
     override suspend fun getWeather(
-        nx: Double,
-        ny: Double,
+        nx: Int,
+        ny: Int,
         baseDateTime: LocalDateTime,
         count: Int,
         date: LocalDate
